Move database module into `lib`
