+++ jobrunner/add_job.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from urllib.parse import urlparse
 
 from jobrunner.create_or_update_jobs import create_or_update_jobs
-from jobrunner.database import find_where
+from jobrunner.lib.database import find_where
 from jobrunner.git import get_sha_from_remote_ref
 from jobrunner.log_utils import configure_logging
 from jobrunner.models import Job
