+++ jobrunner/manage_jobs.py
@@ -17,7 +17,7 @@ import time
 from pathlib import Path
 
 from jobrunner import config, docker
-from jobrunner.database import find_where
+from jobrunner.lib.database import find_where
 from jobrunner.git import checkout_commit
 from jobrunner.models import SavedJobRequest, State, StatusCode
 from jobrunner.path_utils import list_dir_with_ignore_patterns
