+++ jobrunner/run.py
@@ -10,7 +10,7 @@ import sys
 import time
 
 from jobrunner import config
-from jobrunner.database import find_where, select_values, update
+from jobrunner.lib.database import find_where, select_values, update
 from jobrunner.log_utils import configure_logging, set_log_context
 from jobrunner.manage_jobs import (
     JobError,
